Analytics/requirements.txt
flask==1.0.2
Flask-Caching==1.4.0
reportlab==3.5.8
flask_sqlalchemy==2.3.1
alembic==1.0.1
Flask-Babel==0.12.2
flask-login==0.4.1
hashids==1.2.0
MySQL-python==1.2.5
requests==2.20.1